# docs/swagger.yaml
basePath: /api/v1
definitions:
  apperr.Code:
    enum:
    - core/bad_request
    - core/unauthorized
    - core/forbidden
    - core/internal_error
    type: string
    x-enum-varnames:
    - CodeBadRequest
    - CodeUnauthorized
    - CodeForbidden
    - CodeInternal
  apperr.Field:
    enum:
    - request
    type: string
    x-enum-varnames:
    - FieldRequest
  apperr.Rule:
    enum:
    - required
    - too_long
    - cycle
    - max_hierarchy_depth
    - invalid_format
    - too_short
    - duplicate
    - mismatch
    - forbidden
    - invalid_state
    - not_found
    type: string
    x-enum-varnames:
    - RuleRequired
    - RuleTooLong
    - RuleCycle
    - RuleMaxHierarchy
    - RuleInvalidFormat
    - RuleTooShort
    - RuleDuplicate
    - RuleMismatch
    - RuleForbidden
    - RuleInvalidState
    - RuleNotFound
  apperr.Violation:
    properties:
      field:
        $ref: '#/definitions/apperr.Field'
      params:
        additionalProperties: {}
        type: object
      rule:
        $ref: '#/definitions/apperr.Rule'
    type: object
  apperr.appError:
    properties:
      code:
        $ref: '#/definitions/apperr.Code'
      message:
        description: Message for user
        type: string
      violations:
        items:
          $ref: '#/definitions/apperr.Violation'
        type: array
    type: object
...
  http.ChangePasswordInput:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
...
  http.CreateUserInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
  http.UpdateUserInput:
    properties:
      email:
        type: string
      name:
        type: string
    type: object
...
  user.User:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      session_version:
        type: integer
      updated_at:
        type: string
    type: object
info:
  contact: {}
  description: Demo wiki backend in Go
  title: EasyGoDocs API
  version: "1.0"
paths:
...
  /register:
    post:
      consumes:
      - application/json
      description: Creates a new user.
      parameters:
      - description: Create user payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.CreateUserInput'
      responses:
        "201":
          description: Created
        default:
          description: Error
          schema:
            $ref: '#/definitions/apperr.appError'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - users
...
  /users:
    get:
      description: Returns all users. Requires admin role.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.User'
            type: array
        default:
          description: Error
          schema:
            $ref: '#/definitions/apperr.appError'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
  /users/{user_id}:
    delete:
      description: Deletes a user by ID. Requires admin role.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        default:
          description: Error
          schema:
            $ref: '#/definitions/apperr.appError'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      description: Returns a single user by ID. Requires admin role or self.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        default:
          description: Error
          schema:
            $ref: '#/definitions/apperr.appError'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Updates user's basic fields. Requires admin role or self.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Update user payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.UpdateUserInput'
      responses:
        "204":
          description: No Content
        default:
          description: Error
          schema:
            $ref: '#/definitions/apperr.appError'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /users/{user_id}/password:
    post:
      consumes:
      - application/json
      description: Changes password for the specified user (old -> new). Requires
        admin role or self. If admin changes password, old password is not checked.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Change password payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.ChangePasswordInput'
      responses:
        "204":
          description: No Content
        default:
          description: Error
          schema:
            $ref: '#/definitions/apperr.appError'
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - users
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"